Move en-US translations to YAML file
