+++ app/src/lang/available-languages.yaml
@@ -1,0 +1,31 @@
+af-ZA: Afrikaans (South Africa)
+ar-SA: Arabic (Saudi Arabia)
+ca-ES: Catalan (Spain)
+zh-CN: Chinese (Simplified)
+cs-CZ: Czech (Czech Republic)
+da-DK: Danish (Denmark)
+nl-NL: Dutch (Netherlands)
+en-US: English (United States)
+fi-FI: Finnish (Finland)
+fr-FR: French (France)
+de-DE: German (Germany)
+el-GR: Greek (Greece)
+he-IL: Hebrew (Israel)
+hu-HU: Hungarian (Hungary)
+is-IS: Icelandic (Iceland)
+id-ID: Indonesian (Indonesia)
+it-IT: Italian (Italy)
+ja-JP: Japanese (Japan)
+ko-KR: Korean (Korea)
+ms-MY: Malay (Malaysia)
+no-NO: Norwegian (Norway)
+pl-PL: Polish (Poland)
+pt-BR: Portuguese (Brazil)
+pt-PT: Portuguese (Portugal)
+ru-RU: Russian (Russian Federation)
+es-ES: Spanish (Spain)
+es-419: Spanish (Latin America)
+zh-TW: Taiwanese Mandarin (Taiwan)
+tr-TR: Turkish (Turkey)
+uk-UA: Ukrainian (Ukraine)
+vi-VN: Vietnamese (Vietnam)
