--- app/src/lang/en-US/date-format.json
@@ -1,16 +1,0 @@
-{
-    "short": {
-        "year": "numeric",
-        "month": "short",
-        "day": "numeric"
-    },
-    "long": {
-        "year": "numeric",
-        "month": "short",
-        "day": "numeric",
-        "weekday": "short",
-        "hour": "numeric",
-        "minute": "numeric",
-        "hour12": true
-    }
-}
