+++ app/src/lang/en-US/date-formats.yaml
@@ -1,0 +1,13 @@
+short:
+  year: numeric
+  month: short
+  day: numeric
+
+long:
+  year: numeric
+  month: short,
+  day: numeric,
+  weekday: short,
+  hour: numeric,
+  minute: numeric,
+  hour12: true
