--- app/src/lang/en-US/displays.json
@@ -1,106 +1,0 @@
-{
-	"displays": {
-		"boolean": {
-			"boolean": "Boolean",
-			"description": "Display on and off states",
-			"label_on": "Label On",
-			"label_on_placeholder": "Enter an on label...",
-			"label_off": "Label Off",
-			"label_off_placeholder": "Enter an off label...",
-			"icon_on": "Icon On",
-			"icon_off": "Icon Off",
-			"color_on": "Color On",
-			"color_off": "Color Off"
-		},
-		"collection": {
-			"collection": "Collection",
-			"description": "Display a collection",
-			"icon_label": "Show the collection's icon"
-		},
-		"color": {
-			"color": "Color",
-			"description": "Display a colored dot",
-			"default_color": "Default Color",
-			"choices_note": "Set colors relative to the text"
-		},
-		"datetime": {
-			"datetime": "Datetime",
-			"description": "Display values related to time",
-			"format": "Format",
-			"format_note": "The custom format accetps the __[Date Field Symbol Table](https://www.unicode.org/reports/tr35/tr35-dates.html#Date_Field_Symbol_Table)__",
-			"long": "Long",
-			"short": "Short",
-			"relative": "Relative",
-			"relative_label": "Show relative time, eg: 5 minutes ago"
-		},
-		"file": {
-			"file": "File",
-			"description": "Display files"
-		},
-		"filesize": {
-			"filesize": "File Size",
-			"description": "Display the size of a file"
-		},
-		"formatted-value": {
-			"formatted-value": "Formatted Value",
-			"description": "Display a formatted version of the text",
-			"format_title": "Format Title",
-			"format_title_label": "Auto-format casing",
-			"bold_label": "Use bold style"
-		},
-		"formatted-json-value": {
-			"formatted-json-value": "Formatted JSON value",
-			"description": "Display a formatted version of the object"
-		},
-		"icon": {
-			"icon": "Icon",
-			"description": "Display an icon",
-			"filled": "Filled",
-			"filled_label": "Use the filled variant"
-		},
-		"image": {
-			"image": "Image",
-			"description": "Display a tiny image preview",
-			"circle": "Circle",
-			"circle_label": "Display as a circle"
-		},
-		"labels": {
-			"labels": "Labels",
-			"description": "Display either a single or a list of labels",
-			"default_foreground": "Default Foreground",
-			"default_background": "Default Background",
-			"format_label": "Format each label",
-			"show_as_dot": "Show As Dot",
-			"choices_value_placeholder": "Enter a value...",
-			"choices_text_placeholder": "Enter a text..."
-
-		},
-		"mime-type": {
-			"mime-type": "MIME Type",
-			"description": "Show the MIME-Type of a file",
-			"extension_only": "Extension Only",
-			"extension_only_label": "Only show the file extension"
-		},
-		"rating": {
-			"rating": "Rating",
-			"description": "Visualize a number as stars relative to the max-value",
-			"simple": "Simple",
-			"simple_label": "Show stars in a simple format"
-		},
-		"raw": {
-			"raw": "Raw Value"
-		},
-		"related-values": {
-			"related-values": "Related Values",
-			"description": "Display relative values"
-		},
-		"user": {
-			"user": "User",
-			"description": "Display a directus user",
-			"avatar": "Avatar",
-			"name": "Name",
-			"both": "Both",
-			"circle_label": "Show user in a circle"
-		}
-	}
-}
