--- app/src/lang/en-US/interfaces.json
@@ -1,217 +1,0 @@
-{
-	"interfaces": {
-		"checkboxes": {
-			"checkboxes": "Checkboxes",
-			"description": "Choose between multiple options via checkboxes",
-			"allow_other": "Allow Other",
-			"enable_custom_values": "Enable custom values",
-			"show_more": "Show {count} more",
-			"items_shown": "Items Shown"
-		},
-		"code": {
-			"code": "Code",
-			"description": "Write or share code snippets",
-			"line_number": "Line Number",
-			"placeholder": "Enter code here..."
-		},
-		"collection": {
-			"collection": "Collection",
-			"description": "Select between existing collections",
-			"include_system_collections": "Include System Collections"
-		},
-		"collections": {
-			"collections": "Collections",
-			"description": "Select between existing collections",
-			"include_system_collections": "Include System Collections"
-		},
-		"color": {
-			"color": "Color",
-			"description": "Enter or select a color value",
-			"placeholder": "Choose a color...",
-			"preset_colors": "Preset Colors",
-			"preset_colors_add_label": "Add new color...",
-			"name_placeholder": "Enter color name..."
-		},
-		"datetime": {
-			"datetime": "Datetime",
-			"description": "Enter dates and times",
-			"include_seconds": "Include Seconds",
-			"set_to_now": "Set to Now",
-			"use_24": "Use 24-Hour Format"
-		},
-		"display-template": {
-			"display-template": "Display Template",
-			"description": "Mix static text and dynamic field values",
-			"collection_field": "Collection field",
-			"collection_field_not_setup": "The collection field option is misconfigured",
-			"select_a_collection": "Select a Collection"
-		},
-		"divider": {
-			"divider": "Divider",
-			"description": "Label and divide fields into sections",
-			"title_placeholder": "Enter a title...",
-			"inline_title": "Inline Title",
-			"inline_title_label": "Show title inside line"
-		},
-		"dropdown": {
-			"dropdown": "Dropdown",
-			"description": "Select a value from a dropdown",
-			"choices_placeholder": "Add a new choice",
-			"allow_other": "Allow Other",
-			"allow_other_label": "Allow Other Values",
-			"allow_none": "Allow None",
-			"allow_none_label": "Allow No Selection",
-			"choices_name_placeholder": "Enter a name...",
-			"choices_value_placeholder": "Enter a value..."
-		},
-		"dropdown-multiselect": {
-			"dropdown-multiselect": "Dropdown (Multiple)",
-			"description": "Select multiple values from a dropdown"
-		},
-		"file": {
-			"file": "File",
-			"description": "Select or upload a file"
-		},
-		"files": {
-			"files": "Files",
-			"description": "Select or upload multiple files"
-		},
-		"hash": {
-			"hash": "Hash",
-			"description": "Enter a value to be hashed",
-			"masked": "Masked",
-			"masked_label": "Hide the true values"
-		},
-		"icon": {
-			"icon": "Icon",
-			"description": "Select an icon from a dropdown",
-			"search_for_icon": "Search for icon..."
-		},
-		"image": {
-			"image": "Image",
-			"description": "Select or upload an image"
-		},
-		"interface": {
-			"interface": "Interface",
-			"description": "Select an existing interface",
-			"placeholder": "Select an interface..."
-		},
-		"interface-options": {
-			"interface-options": "Interface Options",
-			"description": "A modal for selecting options of an interface"
-		},
-		"many-to-many": {
-			"many-to-many": "Many to Many",
-			"description": "Select multiple related junction items"
-		},
-		"many-to-one": {
-			"many-to-one": "Many to One",
-			"description": "Select a single related item",
-			"display_template": "Display Template"
-		},
-		"markdown": {
-			"markdown": "Markdown",
-			"description": "Enter and preview markdown",
-			"tabbed": "Tabbed",
-			"tabbed_label": "Show preview in separate tab",
-			"edit": "Edit",
-			"preview": "Preview"
-		},
-		"notice": {
-			"notice": "Notice",
-			"description": "Display a short notice",
-			"text": "Enter notice content here..."
-		},
-		"numeric": {
-			"numeric": "Numeric",
-			"description": "Enter a number",
-			"minimum_value": "Minimum Value",
-			"maximum_value": "Maximum Value",
-			"step_interval": "Step Interval"
-		},
-		"one-to-many": {
-			"one-to-many": "One to Many",
-			"description": "Select multiple related items",
-			"readable_fields_copy": "Select the fields that the user can view",
-			"no_collection": "The collection could not be found"
-		},
-		"radio-buttons": {
-			"radio-buttons": "Radio Buttons",
-			"description": "Select one of multiple choices"
-		},
-		"repeater": {
-			"repeater": "Repeater",
-			"description": "Create multiple entries of the same structure",
-			"max_amount": "Maximum Amount",
-			"max_amount_placeholder": "Maximum amount of items...",
-			"edit_fields": "Edit Fields",
-			"add_label": "\"Add New Row\" Label",
-			"field_name_placeholder": "Enter field name...",
-			"field_note_placeholder": "Enter field note..."
-		},
-		"slider": {
-			"slider": "Slider",
-			"description": "Select a number using a slider",
-			"always_show_value": "Always show value"
-		},
-		"slug": {
-			"slug": "Slug",
-			"description": "Enter a word connected with hyphens"
-		},
-		"tags": {
-			"tags": "Tags",
-			"description": "Select or add tags",
-			"whitespace": "Whitespace",
-			"hyphen": "Replace with hyphen",
-			"underscore": "Replace with underscore",
-			"remove": "Remove whitespace",
-			"capitalization": "Capitalization",
-			"uppercase": "Convert Uppercase",
-			"lowercase": "Convert Lowercase",
-			"auto_formatter": "Use Title Auto-Formatter",
-			"alphabetize": "Alphabetize",
-			"alphabetize_label": "Force Alphabetical Order",
-			"add_tags": "Add tags..."
-		},
-		"text-input": {
-			"text-input": "Text Input",
-			"description": "Enter a single line text",
-			"trim": "Trim",
-			"trim_label": "Trim the start and end",
-			"mask": "Masked",
-			"mask_label": "Hide the real value"
-		},
-		"textarea": {
-			"textarea": "Textarea",
-			"description": "Enter multiline plain-text"
-		},
-		"toggle": {
-			"toggle": "Toggle",
-			"description": "Switch between on and off",
-			"label_placeholder": "Enter a label...",
-			"label_default": "Enabled"
-		},
-		"translations": {
-			"translations": "Translations",
-			"display_template": "Display Template",
-			"no_collection": "No Collection"
-		},
-		"user": {
-			"user": "User",
-			"description": "Select an existing directus user",
-			"select_mode": "Select Mode",
-			"modes": {
-				"auto": "Auto",
-				"dropdown": "Dropdown",
-				"modal": "Modal"
-			}
-		},
-		"wysiwyg": {
-			"wysiwyg": "WYSIWYG",
-			"description": "A rich-text editor writing HTML content",
-			"toolbar": "Toolbar",
-			"custom_formats": "Custom Formats",
-			"options_override": "Options Override"
-		}
-	}
-}
